refactor: replace util/ with lib/ and core/ directories
- lib/ for pure algorithms with no VS Code dependency (arrays, lazy, path, range, similarArrayMatch, slugify)
- core/ for VS Code extension infrastructure (dispose, document, file, findFiles, l10n, linkSanitizer, links, resources, workspace)
- topmostLineMonitor moved into features/preview/ as it is only used there
