+++ src/core/dispose.ts
