+++ src/core/document.ts
