+++ src/core/file.ts
