+++ src/core/l10n.ts
