+++ src/core/linkSanitizer.ts
