+++ src/core/resources.ts
