+++ src/features/asciidoctor/asciidoctorAttributesConfig.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode'
-import { findDefaultWorkspaceFolderUri } from '../../util/workspace.js'
+import { findDefaultWorkspaceFolderUri } from '../../core/workspace.js'
 
 export class AsciidoctorAttributesConfig {
   public static getPreviewAttributes(): {} {
