+++ src/features/completion/completionProviders.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode'
-import { disposeAll } from '../../util/dispose.js'
-import { asciidocDocumentSelector } from '../../util/document.js'
+import { disposeAll } from '../../core/dispose.js'
+import { asciidocDocumentSelector } from '../../core/document.js'
 import { AsciidocLoader } from '../asciidoctor/asciidocLoader.js'
 import { BibtexProvider } from './bibtexCompletionProvider.js'
 import { TargetPathCompletionProvider } from './targetPathCompletionProvider.js'
