+++ src/features/completion/xrefCompletionProvider.ts
@@ -1,6 +1,6 @@
 import * as path from 'path'
 import * as vscode from 'vscode'
-import { findFiles } from '../../util/findFiles.js'
+import { findFiles } from '../../core/findFiles.js'
 import { Context, createContext } from './createContext.js'
 
 export const xrefProvider = {
