+++ src/features/documentLinkProvider.ts
@@ -1,10 +1,10 @@
 import * as path from 'path'
 import * as vscode from 'vscode'
 import { OpenDocumentLinkCommand } from '../commands/index.js'
-import { t as l10nT } from '../util/l10n.js'
-import { isSchemeBlacklisted } from '../util/linkSanitizer.js'
-import { getUriForLinkWithKnownExternalScheme } from '../util/links.js'
-import { similarArrayMatch } from '../util/similarArrayMatch.js'
+import { t as l10nT } from '../core/l10n.js'
+import { isSchemeBlacklisted } from '../core/linkSanitizer.js'
+import { getUriForLinkWithKnownExternalScheme } from '../core/links.js'
+import { similarArrayMatch } from '../lib/similarArrayMatch.js'
 import { AsciidocIncludeItemsLoader } from './asciidoctor/asciidocLoader.js'
 
 /**
