+++ src/features/preview/asciidoctorWebViewConverter.ts
@@ -5,11 +5,11 @@ import {
 import * as vscode from 'vscode'
 import * as uri from 'vscode-uri'
 import { AsciidocContributions } from '../../asciidocExtensions.js'
+import { SkinnyTextDocument } from '../../core/document.js'
+import { t as l10nT } from '../../core/l10n.js'
+import { WebviewResourceProvider } from '../../core/resources.js'
+import { getWorkspaceFolder } from '../../core/workspace.js'
 import { AsciidocPreviewSecurityLevel } from '../../security.js'
-import { SkinnyTextDocument } from '../../util/document.js'
-import { t as l10nT } from '../../util/l10n.js'
-import { WebviewResourceProvider } from '../../util/resources.js'
-import { getWorkspaceFolder } from '../../util/workspace.js'
 import { AntoraDocumentContext } from '../antora/antoraContext.js'
 import { AsciidocPreviewConfiguration } from './previewConfig.js'
