+++ src/features/preview/preview.ts
@@ -3,24 +3,24 @@ import * as vscode from 'vscode'
 import * as uri from 'vscode-uri'
 import { AsciidocContributionProvider } from '../../asciidocExtensions.js'
 import { resolveLinkToAsciidocFile } from '../../commands/openDocumentLink.js'
-import { Logger } from '../../logger.js'
-import { Disposable, disposeAll } from '../../util/dispose.js'
-import { isAsciidocFile } from '../../util/file.js'
-import { t as l10nT } from '../../util/l10n.js'
-import { WebviewResourceProvider } from '../../util/resources.js'
-import {
-  AsciidocFileTopmostLineMonitor,
-  getVisibleLine,
-} from '../../util/topmostLineMonitor.js'
+import { Disposable, disposeAll } from '../../core/dispose.js'
+import { isAsciidocFile } from '../../core/file.js'
+import { t as l10nT } from '../../core/l10n.js'
+import { WebviewResourceProvider } from '../../core/resources.js'
 import {
   getWorkspaceFolder,
   getWorkspaceFolders,
-} from '../../util/workspace.js'
+} from '../../core/workspace.js'
+import { Logger } from '../../logger.js'
 import {
   AsciidocPreviewConfiguration,
   AsciidocPreviewConfigurationManager,
 } from './previewConfig.js'
 import { AsciidocContentProvider } from './previewContentProvider.js'
+import {
+  AsciidocFileTopmostLineMonitor,
+  getVisibleLine,
+} from './topmostLineMonitor.js'
 
 export class AsciidocPreview
   extends Disposable
