+++ src/features/preview/previewConfig.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode'
-import { getWorkspaceFolder } from '../../util/workspace.js'
+import { getWorkspaceFolder } from '../../core/workspace.js'
 
 export class AsciidocPreviewConfiguration {
   public static getForResource(resource: vscode.Uri) {
