+++ src/features/preview/topmostLineMonitor.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode'
-import { disposeAll } from './dispose.js'
-import { isAsciidocFile } from './file.js'
+import { disposeAll } from '../../core/dispose.js'
+import { isAsciidocFile } from '../../core/file.js'
 
 export class AsciidocFileTopmostLineMonitor {
   private readonly disposables: vscode.Disposable[] = []
