+++ src/lib/arrays.ts
