+++ src/lib/lazy.ts
