+++ src/lib/path.ts
