--- src/typings/ref.d.ts
@@ -1,1 +1,0 @@
-/// <reference types='@types/node'/>
